Proper Message Handling
The raw []byte message handling that we were doing has been replaced by
the Message/Response pair, where a Message is sent across the channel to
execute on an agent, and one or more Response objects get sent back to
the caller via a returned reply channel.

This allows the caller to get standard error, standard output, and the
exit code via discrete messages, across a single channel.