Enrich Handler capabilities
Handlers now have access to the SSH stderr stream for communicating
non-output output to the Hub.

Handlers now have the ability to return an exit code to the Agent main
loop, which now passes that code back to the Hub via an SSH
`exit-status` request.

Handlers now have the ability to signal to the Agent main loop that it
should shut down and terminate, allowing Agents to be cycled or halted
remotely, based on Hub / Client desires.