Fix agent re-connection bug
There was a goroutine deadlock and/or race relating to the hangup
channel on each connection.  The Hangup() function was nulling out the
channel, which was being reflected across *ALL* consumers, including the
one who was supposed to be reading from the channel.

That's sort of okay though, since there was a deadlock where the
Hangup() call was writing to the channel, and then expecting the other
part of its originating `select ...` block to read from.  This is now a
buffered channel to get around that bit.  Since we only ever write to
the channel once, this should not cause any issues.

Fixes #7 and #9.